Refactor the error handling system with error_chain.
In order to return all the existings errors and avoiding any unwrap a
more improve error handling was needed. The `error_chain` package is one
of the most used and battle-tested lib for this job.

It allow:
- Wrapping any kind of internal/external error
- Error chaining
- Different level of debuging
- A generic way to handle errors any errors

This commit bring a massive code change but few breaking change in the
current API:

- The `new` method return now a `Result<Box<ExchangeApi>>`
- The `new_from_file` method return now a `Result<Box<ExchangeApi>>`